Rename vector test module
